Add tail-call elimination
The tail of a procedure or cond returns
a special tailCall object which captures
the expression and environment
instead of evaluating it.

`eval` is updated to contain a loop,
where the expression and environment are updated
whena tail call is returned.

One exception to tail-call elimination is withErrorHandler.
It is implemented with panic and defer,
which means we need a go stack frame to be maintained,
so tail calls are forced.